scripts/requirements.txt
requests-html
mysql-connector-python